Add filtering functionality
I originally wanted to make the dropdown searchable live.
That proved to be trickier than expected, so I decided to just go
with filtering for now. I ironed out most of the major bugs and organized
things a bit. I would consider this to be good enough to be used at this point.
Fun and interesting project, I must say. Timezones are more than they appear.